Fix location of cjdns peer injection script
It was being dropped by Ansible in user binaries path while it's a system
binary (it modifies a system config file).

Not a big problem since it's run directly by Ansible, but in some nodes the
good location is used in cjdns' unit file and the file might not be there.

A maintenance task removes the file from its old location.