Install a systemd service to run the dCDN client
It uses the dCDN proxy port and the (new) network hash salt options to start
the dCDN client attached to a particular network.

The proxy port is cosed to the exterior at the firewall while the client is
running.

The client does not run when the testbed is in dormant mode.

The client still needs to be built.